[Refactor]#141 KakaoGallerySharedPreferences를 data.local 패키지로 이동
 - clean architecture에 의하면 DB(Local), Api관련 코드는 data layer에 위치
