+++ app/src/main/java/com/gallery/kakaogallery/data/dao/SaveImageDaoImpl.kt
@@ -1,6 +1,6 @@
 package com.gallery.kakaogallery.data.dao
 
-import com.gallery.kakaogallery.KakaoGallerySharedPreferences
+import com.gallery.kakaogallery.data.local.KakaoGallerySharedPreferences
 import com.gallery.kakaogallery.data.entity.local.ImageEntity
 import com.gallery.kakaogallery.domain.model.SearchImageModel
 import com.google.gson.Gson
